refactor: 불필요한 코드 삭제, int 형으로 정의한 금액 변수 long 형으로 수정

+++ src/Hotel.java
@@ -64,7 +64,7 @@ public class Hotel {
         return impossibleList;
     }
 
-    public Guest findOrCreateGuest(String name, String phone, int cash) {
+    public Guest findOrCreateGuest(String name, String phone, long cash) {
         for(Guest guest : guestsList){
             if(guest.getName().equals(name)){
                 if(guest.getPhone().equals(phone)){
